Create Utilities cog, slim cogs to just commands.stuff
<ChangeDetails>
    <ChangeDescription>Create Utilities cog, slim cogs to just commands.stuff</ChangeDescription>
    <ChangeReasoning>Streamlines cogs</ChangeReasoning>
</ChangeDetails>
